fixed the NaN bug...Ithink

+++ src/three_state_coherence.c
@@ -708,7 +708,7 @@ simulation_data *run_ts_coherence_simulation (int SIMULATION_TYPE, DESIGN *desig
   if ((end_time = time (NULL)) < 0)
     fprintf (stderr, "Could not get end time\n");
 	
-  command_history_message ("Total simulation time: %g s\n", (double)(end_time - start_time));
+  command_history_message ("Total simulation time: %g s\n", (long double)(end_time - start_time));
   set_progress_bar_visible (FALSE) ;
   return sim_data;
   }//run_ts_coherence
